pyproject.toml
[project]
name = "remoteled"
version = "1.0.0"
description = "Control a physical LED over BLE using Raspberry Pi and Android"
requires-python = ">=3.11"
dependencies = [
    "ecdsa>=0.19.1",
    "fastapi>=0.118.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "uvicorn[standard]>=0.30.0",
    "bleak>=0.22.0",
    "stripe>=13.0.1",
    "bluezero>=0.9.1; sys_platform == 'linux'",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "cryptography>=41.0.7",
    "firebase-admin>=7.1.0",
]

[project.optional-dependencies]
pi = [
    "RPi.GPIO>=0.7.1",
    "bluezero>=0.9.1",
    "dbus-python>=1.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pi/python"]
